requirements.txt
absl-py==2.3.0
accelerate==0.30.1
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
bangla==0.0.5
bark==0.1.5
basicsr==1.4.2
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.38.46
botocore==1.38.46
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
ctranslate2==4.6.0
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
dataclasses-json==0.6.7
dateparser==1.1.8
decorator==5.2.1
diffusers==0.34.0
docopt==0.6.2
edge-tts==7.0.2
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.3.0
facexlib==0.3.0
fastapi==0.115.14
ffmpy==0.6.0
filelock==3.18.0
filterpy==1.4.5
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
funcy==2.0
future==1.0.0
g2pkk==0.1.2
gdown==5.2.0
gfpgan==1.3.8
gradio==5.35.0
gradio_client==1.10.4
greenlet==3.2.3
groovy==0.1.2
grpcio==1.73.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
inflect==7.5.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.4
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
lmdb==1.6.2
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.4
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.0
murmurhash==1.0.13
mypy_extensions==1.1.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.26.4
onnxruntime==1.22.0
openai-whisper==20250625
opencv-python==4.11.0.86
orjson==3.10.18
packaging==24.2
pandas==1.5.3
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.54.0
pyreadline3==3.5.4
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
ruff==0.12.1
s3transfer==0.13.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.11.4
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
srt==3.5.3
starlette==0.46.2
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.14.0
tabulate==0.9.0
tb-nightly==2.20.0a20250627
tbb==2021.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.67.1
trainer==0.0.36
transformers==4.53.0
TTS==0.22.0
typeguard==4.4.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.34.3
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yacs==0.1.8
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
